refactor: split TRBs into three files
