+++ kernel/src/device/pci/xhci/exchanger/command.rs
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-use super::super::ring::{command, trb::CommandComplete};
+use super::super::ring::{command, event::trb::CommandCompletion};
 use alloc::{collections::BTreeMap, rc::Rc};
 use core::{
     cell::RefCell,
@@ -52,13 +52,13 @@ impl Sender {
             .add_entry(addr_to_trb, self.waker.clone());
     }
 
-    async fn get_trb(&mut self, addr_to_trb: PhysAddr) -> CommandComplete {
+    async fn get_trb(&mut self, addr_to_trb: PhysAddr) -> CommandCompletion {
         ReceiveFuture::new(addr_to_trb, self.receiver.clone(), self.waker.clone()).await
     }
 }
 
 pub struct Receiver {
-    trbs: BTreeMap<PhysAddr, Option<CommandComplete>>,
+    trbs: BTreeMap<PhysAddr, Option<CommandCompletion>>,
     wakers: BTreeMap<PhysAddr, Rc<RefCell<AtomicWaker>>>,
 }
 impl Receiver {
@@ -88,21 +88,21 @@ impl Receiver {
         Ok(())
     }
 
-    pub fn receive(&mut self, trb: CommandComplete) {
+    pub fn receive(&mut self, trb: CommandCompletion) {
         if let Err(e) = self.insert_trb_and_wake_runner(trb) {
             panic!("Failed to receive a command completion trb: {:?}", e);
         }
     }
 
-    fn insert_trb_and_wake_runner(&mut self, trb: CommandComplete) -> Result<(), Error> {
-        let addr_to_trb = PhysAddr::new(trb.addr_to_command_trb());
+    fn insert_trb_and_wake_runner(&mut self, trb: CommandCompletion) -> Result<(), Error> {
+        let addr_to_trb = PhysAddr::new(trb.trb_addr());
         self.insert_trb(trb)?;
         self.wake_runner(addr_to_trb)?;
         Ok(())
     }
 
-    fn insert_trb(&mut self, trb: CommandComplete) -> Result<(), Error> {
-        let addr_to_trb = PhysAddr::new(trb.addr_to_command_trb());
+    fn insert_trb(&mut self, trb: CommandCompletion) -> Result<(), Error> {
+        let addr_to_trb = PhysAddr::new(trb.trb_addr());
         *self
             .trbs
             .get_mut(&addr_to_trb)
@@ -126,7 +126,7 @@ impl Receiver {
         }
     }
 
-    fn remove_entry(&mut self, addr_to_trb: PhysAddr) -> Option<CommandComplete> {
+    fn remove_entry(&mut self, addr_to_trb: PhysAddr) -> Option<CommandCompletion> {
         match self.trbs.remove(&addr_to_trb) {
             Some(trb) => trb,
             None => panic!("No such receiver with TRB address: {:?}", addr_to_trb),
@@ -159,7 +159,7 @@ impl ReceiveFuture {
     }
 }
 impl Future for ReceiveFuture {
-    type Output = CommandComplete;
+    type Output = CommandCompletion;
 
     fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
         let waker = self.waker.clone();
