+++ kernel/src/device/pci/xhci/ring/mod.rs
@@ -4,7 +4,6 @@ pub mod command;
 pub mod event;
 mod raw;
 pub mod transfer;
-pub mod trb;
 
 #[derive(Copy, Clone, PartialOrd, Ord, PartialEq, Eq)]
 pub struct CycleBit(bool);
